+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## HEAD
 
+- App#directory is deprecated (https://github.com/heroku/hatchet/pull/135)
 - Empty string returns from App#run now trigger retries (https://github.com/heroku/hatchet/pull/132)
 - Annotate rspec expectation failures inside of deploy blocks with hatchet debug information (https://github.com/heroku/hatchet/pull/136)
 - Hatchet#new raises a helpful error when no source code location is provided (https://github.com/heroku/hatchet/pull/134)